# Add container caching for faster CI builds