--- a/game_tests/final_test.cr
+++ b/game_tests/final_test.cr
@@ -4,7 +4,7 @@ require "json"
 # Complete test of the game system according to requirements
 def websocket_handshake(socket, host, port)
   key = "dGhlIHNhbXBsZSBub25jZQ=="
-  
+
   request = String.build do |str|
     str << "GET / HTTP/1.1\r\n"
     str << "Host: #{host}:#{port}\r\n"
@@ -14,47 +14,47 @@ def websocket_handshake(socket, host, port)
     str << "Sec-WebSocket-Version: 13\r\n"
     str << "\r\n"
   end
-  
+
   socket.print(request)
-  
+
   response = ""
   while line = socket.gets
     response += line
     break if line.strip.empty?
   end
-  
+
   response.includes?("101 Switching Protocols")
 end
 
 def encode_frame(payload : String)
   bytes = payload.to_slice
   frame = IO::Memory.new
-  
+
   frame.write_byte(0x81_u8)
-  
+
   if bytes.size < 126
     frame.write_byte((bytes.size | 0x80).to_u8)
   else
     frame.write_byte(0xFE_u8)
     frame.write_bytes(bytes.size.to_u16, IO::ByteFormat::BigEndian)
   end
-  
+
   mask = Bytes[0x12, 0x34, 0x56, 0x78]
   frame.write(mask)
-  
+
   bytes.each_with_index do |byte, i|
     frame.write_byte((byte ^ mask[i % 4]).to_u8)
   end
-  
+
   frame.to_slice
 end
 
 def decode_frame(socket)
   first_byte = socket.read_byte || return ""
   second_byte = socket.read_byte || return ""
-  
+
   payload_len = (second_byte & 0x7F).to_i
-  
+
   if payload_len == 126
     len_bytes = Bytes.new(2)
     socket.read_fully(len_bytes)
@@ -64,7 +64,7 @@ def decode_frame(socket)
     socket.read_fully(len_bytes)
     payload_len = (len_bytes[4].to_u32 << 24) | (len_bytes[5].to_u32 << 16) | (len_bytes[6].to_u32 << 8) | len_bytes[7].to_u32
   end
-  
+
   payload = Bytes.new(payload_len)
   socket.read_fully(payload)
   String.new(payload)
@@ -86,53 +86,53 @@ puts "6) Validate game state files\n"
 
 begin
   socket = TCPSocket.new("localhost", 3000)
-  
+
   if websocket_handshake(socket, "localhost", 3000)
     puts "✅ 1) Connected client to server"
-    
+
     # Send initial empty message
     socket.write(encode_frame(""))
-    
+
     # Read welcome message and extract UUID
     welcome = decode_frame(socket)
     puts "   #{welcome}"
-    
+
     uuid = ""
     if welcome.includes?("UUID:")
       uuid = welcome.split("UUID: ")[1].split(",")[0]
     end
-    
+
     # Create client state file (simulating client behavior)
     client_state = {"uuid" => uuid, "name" => nil}
     File.write("client/game_state.json", client_state.to_json)
-    
+
     # 2) /say Hello?
     puts "\n📤 2) Issuing command: /say Hello?"
     socket.write(encode_frame("/say Hello?"))
     response = decode_frame(socket)
     puts "   Server response: #{response}"
-    
+
     # 3) /cast Illuminate 0,0
     puts "\n📤 3) Issuing command: /cast Illuminate 0,0"
     socket.write(encode_frame("/cast Illuminate 0,0"))
     response = decode_frame(socket)
     puts "   Server response: #{response}"
-    
+
     # 4) /alias Ford Prefect
     puts "\n📤 4) Issuing command: /alias Ford Prefect"
     socket.write(encode_frame("/alias Ford Prefect"))
     response = decode_frame(socket)
     puts "   Server response: #{response}"
-    
+
     # Update client state with new name
     client_state = {"uuid" => uuid, "name" => "Ford Prefect"}
     File.write("client/game_state.json", client_state.to_json)
-    
+
     # Test /witness command
     puts "\n📤 Testing /witness command:"
     socket.write(encode_frame("/witness"))
     witness_response = decode_frame(socket)
-    
+
     if witness_response.starts_with?("{") && witness_response.includes?("log")
       puts "✅ /witness command returns properly formatted JSON:"
       puts "="*50
@@ -142,15 +142,13 @@ begin
     else
       puts "⚠️  /witness response: #{witness_response}"
     end
-    
+
     puts "\n🔌 5) Disconnecting client..."
     socket.close
-    
   else
     puts "❌ Failed to connect to server"
     exit 1
   end
-  
 rescue ex
   puts "❌ Test failed: #{ex.message}"
   exit 1
@@ -163,24 +161,24 @@ puts "=================================="
 if File.exists?("server/game_state.json")
   puts "✅ Server game state file exists"
   server_content = File.read("server/game_state.json")
-  
+
   begin
     server_json = JSON.parse(server_content)
     puts "✅ Server state is valid JSON"
-    
+
     if server_json["log"]? && server_json["players"]?
       puts "✅ Server state has correct structure (log, players)"
-      
+
       log = server_json["log"].as_a
       puts "   📊 Game log contains #{log.size} coordinate entries"
-      
+
       players = server_json["players"].as_h
       puts "   👥 Players tracked: #{players.size}"
-      
+
       if players.any? { |_, p| p.as_h["name"]?.try(&.as_s) == "Ford Prefect" }
         puts "✅ Player name correctly set to 'Ford Prefect'"
       end
-      
+
       # Check for spells cast
       spells_cast = 0
       log.each do |entry|
@@ -189,11 +187,9 @@ if File.exists?("server/game_state.json")
         end
       end
       puts "   ⚡ Spells cast recorded: #{spells_cast}"
-      
     else
       puts "⚠️  Server state missing required fields"
     end
-    
   rescue
     puts "⚠️  Server state is not valid JSON"
   end
@@ -207,28 +203,26 @@ puts ""
 if File.exists?("client/game_state.json")
   puts "✅ Client game state file exists"
   client_content = File.read("client/game_state.json")
-  
+
   begin
     client_json = JSON.parse(client_content)
     puts "✅ Client state is valid JSON"
-    
+
     if client_json["uuid"]? && client_json["name"]?
       puts "✅ Client state has correct structure (uuid, name)"
-      
+
       uuid = client_json["uuid"]?.try(&.as_s)
       name = client_json["name"]?.try(&.as_s)
-      
+
       puts "   🔑 UUID: #{uuid}"
       puts "   👤 Name: #{name}"
-      
+
       if name == "Ford Prefect"
         puts "✅ Client name correctly set to 'Ford Prefect'"
       end
-      
     else
       puts "⚠️  Client state missing required fields"
     end
-    
   rescue
     puts "⚠️  Client state is not valid JSON"
   end