## Summary

- Add automated container building and pushing to Docker Hub on main branch pushes
- Update all CI workflows to use cached containers for significantly faster builds
- Configure Docker Compose with cache_from settings for optimal caching
- Migrate from GitHub Container Registry to Docker Hub for better performance

## Benefits

- **Faster CI/CD pipelines**: Reuse precompiled Crystal binaries instead of compiling from scratch
- **Multi-platform support**: Build for both linux/amd64 and linux/arm64
- **Automatic cache invalidation**: Cache updates when dependencies or source code changes
- **Reduced resource usage**: Less CPU time and bandwidth usage in CI
- **Better developer experience**: Faster feedback on pull requests
- **Improved reliability**: Docker Hub's proven infrastructure and availability

## Technical Details

### New Workflow: `build-and-push.yml`
- Triggers on pushes to main branch
- Builds and pushes containers to `docker.io/robcole/yolo.cr`
- Uses GitHub Actions cache and multi-platform building
- Tags containers with branch name, SHA, and latest

### Updated Workflows
- **CI workflow**: Pre-builds images once, tests multiple times using Docker Hub
- **Integration tests**: Uses pre-compiled binaries for faster, more reliable testing
- **Docker Compose**: Configured with `cache_from` for all services

### Container Strategy
1. **Registry**: Docker Hub (`docker.io`) with DOCKER_HUB_TOKEN authentication
2. **Pre-built binaries**: Build once in separate job, test multiple times
3. **Optimized health checks**: 5s intervals, 10 retries, 20s start period
4. **Cache layers**: GitHub Actions cache + Docker Hub registry cache
5. **Unique tagging**: `ci-{sha}` and `pr-{number}` for isolation

## Performance Optimizations

- ✅ **Split build and test phases**: Heavy compilation happens once
- ✅ **Pre-built binary strategy**: No Crystal compilation during testing
- ✅ **Direct docker run commands**: Faster than docker-compose overhead
- ✅ **Optimized health check timing**: Reduced from 45s to 20s start period
- ✅ **Parallel test execution**: Lint and connectivity tests run simultaneously

## Fixes Applied

- ✅ **Fixed `unknown flag: --cache-from` error**: Removed unsupported flags from docker compose build
- ✅ **Fixed health check failures**: Added curl to runtime image, replaced wget commands
- ✅ **Fixed resource constraints**: Pre-built binaries eliminate compilation load during testing
- ✅ **Fixed container registry**: Migrated to Docker Hub for better performance
- ✅ **Fixed Crystal formatting**: Applied consistent formatting to all source files
- ✅ **Fixed YAML formatting**: Removed trailing whitespace, added proper newlines

## Test Plan

- [x] Build Docker containers locally with caching
- [x] Verify cache_from configuration works in docker-compose
- [x] Fix CI workflow compatibility issues 
- [x] Fix health check dependencies and timing
- [x] Optimize workflows with pre-built binary caching
- [x] Migrate to Docker Hub registry
- [x] Apply consistent Crystal and YAML formatting
- [ ] Monitor CI build times after merge (expect 50-80% reduction)

## How It Works

1. **Build Phase**: Compile Crystal binaries and push to Docker Hub
2. **Test Phase**: Pull pre-built images and run fast tests
3. **Health checks**: Optimized timing for immediate startup with pre-built binaries
4. **Caching**: Multi-layer strategy with GitHub Actions + Docker Hub
5. **Authentication**: Uses DOCKER_HUB_TOKEN for reliable registry access

## Migration Benefits

**Docker Hub vs GHCR:**
- Better availability and performance
- No GitHub package permissions required
- Native Docker ecosystem integration
- Proven infrastructure for container distribution
- Configured DOCKER_HUB_TOKEN authentication

This change will significantly improve CI performance while maintaining the same test coverage and quality checks. The optimized workflows should resolve resource constraint issues and provide faster, more reliable testing.